added some simple polling for new issues
This allows users of the client to subscribe to events for new
or updated issues.

This is an extremely simple and naive implementation. Because of
an apparent error in the way Redmine parses dates that have times
we are not passing an exact date/time combo yet. We are simply
getting all issues from the past day and the current day.